Add wPlcException Type
Added lots of error checking to PLC methods. Any time a PLC operation is
conducted, the result is checked for an error code. If an error is
encountered, a wPLCException will be thrown, and a message will be shown
to the user with the error code. These codes should be expanded upon at a
later date with friendlier error messages.

All main controls now enable/disable as appropriate.